requirement.txt
streamlit>=1.28.0
pillow>=10.0.0
matplotlib>=3.7.0